requirements.txt
altair==4.2.2
asgiref==3.6.0
attrs==22.2.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
Django==4.1.7
entrypoints==0.4
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
image==1.5.33
importlib-metadata==6.1.0
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.3
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rich==13.3.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
sqlparse==0.4.3
streamlit==1.20.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
zipp==3.15.0
